Working in the templates of analytical_solutions
