+++ src/analytical_solutions/noh_solution/main.cpp
@@ -48,8 +48,8 @@
 using namespace std;
 using namespace sphexa;
 
-using Real    = double;
-using KeyType = uint64_t;
+using T = double;
+using I = uint64_t;
 
 void printHelp(char* binName);
 
@@ -65,28 +65,28 @@ int main(int argc, char** argv)
 
     // Get command line parameters
     const bool   only_sol  = parser.exists("--only_solution") ? true : false;
-    const double time      = parser.getDouble("--time", 0.);
-    const size_t nParts    = parser.getInt("--nParts", 0);
+    const T      time      = parser.getDouble("--time", 0.);
+    const I      nParts    = parser.getInt("--nParts", 0);
     const string inputFile = parser.getString("--input", "./dump_noh0.txt");
     const string outDir    = parser.getString("--outDir", "./");
     const bool   complete  = parser.exists("--complete") ? true : false;
 
     // Initialize vector size
-    vector<double>     x(nParts);
-    vector<double>     y(nParts);
-    vector<double>     z(nParts);
-    vector<double>     vx(nParts);
-    vector<double>     vy(nParts);
-    vector<double>     vz(nParts);
-    vector<double>     h(nParts);
-    vector<double>     rho(nParts);
-    vector<double>     u(nParts);
-    vector<double>     p(nParts);
-    vector<double>     cs(nParts);
-    vector<double>     Px(nParts);
-    vector<double>     Py(nParts);
-    vector<double>     Pz(nParts);
-    vector<ParticleIO> vSim(nParts);
+    vector<T>                x(nParts);
+    vector<T>                y(nParts);
+    vector<T>                z(nParts);
+    vector<T>                vx(nParts);
+    vector<T>                vy(nParts);
+    vector<T>                vz(nParts);
+    vector<T>                h(nParts);
+    vector<T>                rho(nParts);
+    vector<T>                u(nParts);
+    vector<T>                p(nParts);
+    vector<T>                cs(nParts);
+    vector<T>                Px(nParts);
+    vector<T>                Py(nParts);
+    vector<T>                Pz(nParts);
+    vector<ParticleIO<T, I>> vSim(nParts);
 
     if (!only_sol)
     {
@@ -102,18 +102,18 @@ int main(int argc, char** argv)
         }
 
         // Load particles data
-        NohFileData::readData3D(inputFile, nParts, x, y, z, vx, vy, vz, h, rho, u, p, cs, Px, Py, Pz);
+        NohFileData<T, I>::readData3D(inputFile, nParts, x, y, z, vx, vy, vz, h, rho, u, p, cs, Px, Py, Pz);
 
         // Calculate radius, velocity and sort particle data by radius
-        for (size_t i = 0; i < nParts; i++)
+        for (I i = 0; i < nParts; i++)
         {
-            double r   = sqrt((x[i] * x[i]) + (y[i] * y[i]) + (z[i] * z[i]));
-            double vel = sqrt((vx[i] * vx[i]) + (vy[i] * vy[i]) + (vz[i] * vz[i]));
+            T r   = sqrt((x[i] * x[i]) + (y[i] * y[i]) + (z[i] * z[i]));
+            T vel = sqrt((vx[i] * vx[i]) + (vy[i] * vy[i]) + (vz[i] * vz[i]));
 
             vSim[i] = {
                 i, r, vel, x[i], y[i], z[i], vx[i], vy[i], vz[i], h[i], rho[i], u[i], p[i], cs[i], Px[i], Py[i], Pz[i]};
         }
-        sort(vSim.begin(), vSim.end(), ParticleIO::cmp());
+        sort(vSim.begin(), vSim.end(), ParticleIO<T, I>::cmp());
     }
 
     // Get time without rounding
@@ -122,26 +122,26 @@ int main(int argc, char** argv)
     string time_str = time_long.str();
 
     // Calculate and write theoretical solution profile in one dimension    // time = 0.6s
-    const size_t dim     = NohDataGenerator<Real, KeyType>::dim;   // 3
-    const double r0      = NohDataGenerator<Real, KeyType>::r0;    // .0
-    const double r1      = NohDataGenerator<Real, KeyType>::r1;    // .1
-    const double gamma   = NohDataGenerator<Real, KeyType>::gamma; // 5./3.
-    const double rho0    = NohDataGenerator<Real, KeyType>::rho0;  // 1.
-    const double u0      = NohDataGenerator<Real, KeyType>::u0;    // 0.
-    const double p0      = NohDataGenerator<Real, KeyType>::p0;    // 0.
-    const double vr0     = NohDataGenerator<Real, KeyType>::vr0;   // -1.
-    const double cs0     = NohDataGenerator<Real, KeyType>::cs0;   // 0.
+    const size_t dim     = NohDataGenerator<T, I>::dim;   // 3
+    const double r0      = NohDataGenerator<T, I>::r0;    // .0
+    const double r1      = NohDataGenerator<T, I>::r1;    // .1
+    const double gamma   = NohDataGenerator<T, I>::gamma; // 5./3.
+    const double rho0    = NohDataGenerator<T, I>::rho0;  // 1.
+    const double u0      = NohDataGenerator<T, I>::u0;    // 0.
+    const double p0      = NohDataGenerator<T, I>::p0;    // 0.
+    const double vr0     = NohDataGenerator<T, I>::vr0;   // -1.
+    const double cs0     = NohDataGenerator<T, I>::cs0;   // 0.
     const string solFile = outDir + "noh_solution_" + time_str + ".dat";
 
     // Set the positions for calculate the solution
-    vector<double> rSol;
-    size_t         nSteps = 1000;
+    vector<T> rSol;
+    I         nSteps = 1000;
     if (only_sol || !complete)
     {
-        const double rMax  = r1; // 2. * r1;
-        const double rStep = (rMax - r0) / nSteps;
+        const T rMax  = r1; // 2. * r1;
+        const T rStep = (rMax - r0) / nSteps;
 
-        for (size_t i = 0; i < nSteps; i++)
+        for (I i = 0; i < nSteps; i++)
         {
             rSol.push_back(r0 + (0.5 * rStep) + (i * rStep));
         }
@@ -150,14 +150,14 @@ int main(int argc, char** argv)
     {
         nSteps = nParts;
 
-        for (size_t i = 0; i < nSteps; i++)
+        for (I i = 0; i < nSteps; i++)
         {
             rSol.push_back(vSim[i].r);
         }
     }
 
     // Calculate Sedov solution
-    NohSolution::create(rSol, dim, nSteps, time, gamma, rho0, u0, p0, vr0, cs0, solFile);
+    NohSolution<T, I>::create(rSol, dim, nSteps, time, gamma, rho0, u0, p0, vr0, cs0, solFile);
 
     // Write Info: Output files and colums.
     cout << "\nExecuted successfully.\n";
@@ -167,12 +167,12 @@ int main(int argc, char** argv)
     {
         // Write 1D simulation solution to compare with the theoretical solution
         const string simFile = outDir + "noh_simulation_" + time_str + ".dat";
-        NohFileData::writeParticle1D(nParts, vSim, simFile);
+        NohFileData<T, I>::writeParticle1D(nParts, vSim, simFile);
         cout << "Simulation file: '" << simFile << "'";
     }
 
     cout << "\nColumns:\n";
-    NohFileData::writeColumns1D(cout);
+    NohFileData<T, I>::writeColumns1D(cout);
     cout << endl;
 
     exit(EXIT_SUCCESS);
