+++ src/analytical_solutions/sedov_solution/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(SOURCES sedov_io.cpp sedov_solution.cpp main.cpp)
+set(SOURCES main.cpp)
 add_executable(sedov_solution ${SOURCES})
 target_include_directories(sedov_solution PRIVATE ${PROJECT_SOURCE_DIR}/include)
 target_include_directories(sedov_solution PRIVATE ${PROJECT_SOURCE_DIR}/src)
