+++ src/sedov/sedov.cpp
@@ -11,10 +11,10 @@
 #include "cstone/domain/domain.hpp"
 #include "sphexa.hpp"
 #include "sph/findNeighborsSfc.hpp"
-#include "common/test_file_writer.hpp"
 #include "timer.hpp"
 #include "propagator.hpp"
-#include "insitu_viz.h"
+#include "../common/test_file_writer.hpp"
+#include "../insitu_viz.h"
 
 #include "SedovDataGenerator.hpp"
